Smarter `git out`.
Shows commits not merged into next closest branch, not just
origin/master, so can tell you a lot of things:

* what have I done on this topic branch?
* what haven't I pushed to GitHub?
* what haven't I deployed to Heroku?

But I'm just now trying it out, so I bet this is going to show way too
many commits when on an old topic branch that's not been rebased, and
when HEAD == master == origin/master == heroku/master.

I may be doing too much, but for the life of me I can't figure out how
to say "the closest parent branch" in a regular revision parameter to
git log. (I don't understand @{upstream} well enough to know why it's
not working for me.)