Trying out ack.vim.
I've been slow to want to start doing command-line-ish things in Vim,
but perhaps I should reconsider.

It is generally the case that I'm acking in order to open a file,
though, so this seems like a good plunge to take.

Tim Pope's unimpaired provides nice mappings, [q and ]q, for moving
around in the quickfix list, so I expect I'll be getting those under my
fingers soon.